# config/platform/session_api_v1.openapi.json
{
  "components": {
    "schemas": {
      "AnalysisRunResponse": {
        "properties": {
          "bundle_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Fingerprint"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "core_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Core Version"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "processing_profile": {
            "title": "Processing Profile",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "session_id",
          "status",
          "processing_profile",
          "created_at"
        ],
        "title": "AnalysisRunResponse",
        "type": "object"
      },
      "AnalysisRunSummary": {
        "properties": {
          "bundle_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Fingerprint"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "processing_profile": {
            "title": "Processing Profile",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "processing_profile"
        ],
        "title": "AnalysisRunSummary",
        "type": "object"
      },
      "ArtifactResponse": {
        "properties": {
          "analysis_run_id": {
            "format": "uuid",
            "title": "Analysis Run Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "analysis_run_id",
          "kind",
          "object_key",
          "media_type",
          "size_bytes",
          "sha256",
          "created_at"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
      "ErrorPayload": {
        "properties": {
          "code": {
            "examples": [
              "SESSION_NOT_FOUND"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "examples": [
              "session not found"
            ],
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "examples": [
              "2f9e4f25-9d45-4e04-a5e7-8dd3b6c2d310"
            ],
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorPayload",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "MediaResponse": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "integrity_status": {
            "title": "Integrity Status",
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "download_url",
          "expires_at",
          "content_type",
          "size_bytes",
          "sha256",
          "integrity_status"
        ],
        "title": "MediaResponse",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "bundle_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Fingerprint"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_analysis_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisRunSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "processing_profile": {
            "title": "Processing Profile",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "surface": {
            "title": "Surface",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "video": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "processing_profile",
          "surface",
          "created_at",
          "updated_at"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "SessionStatus": {
        "enum": [
          "DRAFT",
          "AWAITING_UPLOAD",
          "UPLOADING",
          "UPLOADED",
          "QUEUED",
          "PROCESSING",
          "COMPLETE",
          "PARTIAL",
          "FAILED"
        ],
        "title": "SessionStatus",
        "type": "string"
      },
      "UploadComplete": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "size_bytes",
          "content_type"
        ],
        "title": "UploadComplete",
        "type": "object"
      },
      "UploadCompleteResponse": {
        "properties": {
          "integrity_status": {
            "title": "Integrity Status",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "video_id": {
            "format": "uuid",
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "status",
          "integrity_status"
        ],
        "title": "UploadCompleteResponse",
        "type": "object"
      },
      "UploadInitiate": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "display_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "display_name",
          "content_type",
          "size_bytes"
        ],
        "title": "UploadInitiate",
        "type": "object"
      },
      "UploadResponse": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "required_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Required Headers",
            "type": "object"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          },
          "video_id": {
            "format": "uuid",
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "object_key",
          "upload_url",
          "method",
          "required_headers",
          "expires_at"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
      "VideoSummary": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "integrity_status": {
            "title": "Integrity Status",
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display_name",
          "content_type",
          "size_bytes",
          "sha256",
          "integrity_status"
        ],
        "title": "VideoSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local-first Session API V1 candidate",
    "title": "TennisAI Session Platform",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions/{session_id}": {
      "get": {
        "description": "Return session metadata and the source-video summary when available.",
        "operationId": "getSession",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "status": "DRAFT"
                },
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Session metadata"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get a session",
        "tags": [
          "Sessions"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/analysis-runs": {
      "get": {
        "description": "List analysis-run metadata for a session.",
        "operationId": "listAnalysisRuns",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [],
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRunResponse"
                  },
                  "title": "Response Listanalysisruns",
                  "type": "array"
                }
              }
            },
            "description": "Analysis runs"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List analysis runs",
        "tags": [
          "Analysis Runs"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/artifacts": {
      "get": {
        "description": "List analysis bundle artifacts linked to the session's analysis runs.",
        "operationId": "listArtifacts",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [],
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactResponse"
                  },
                  "title": "Response Listartifacts",
                  "type": "array"
                }
              }
            },
            "description": "Artifacts"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List session artifacts",
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/media": {
      "get": {
        "description": "Return a short-lived presigned download URL for the session source video.",
        "operationId": "getSessionMedia",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "download_url": "https://storage.example/object",
                  "integrity_status": "STORAGE_VERIFIED"
                },
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Presigned media URL"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get session media",
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/uploads": {
      "post": {
        "description": "Validate source-video metadata and return a short-lived presigned PUT URL.",
        "operationId": "initiateVideoUpload",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "mp4": {
                  "summary": "MP4 source video",
                  "value": {
                    "content_type": "video/mp4",
                    "display_name": "rally.mp4",
                    "size_bytes": 1048576
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/UploadInitiate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "method": "PUT",
                  "object_key": "sessions/..."
                },
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Presigned upload URL"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Initiate a video upload",
        "tags": [
          "Uploads"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/uploads/{video_id}/complete": {
      "post": {
        "description": "HEAD the object in MinIO/S3 and mark it STORAGE_VERIFIED when metadata matches.",
        "operationId": "completeVideoUpload",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "uploaded": {
                  "summary": "Uploaded object metadata",
                  "value": {
                    "content_type": "video/mp4",
                    "size_bytes": 1048576
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/UploadComplete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "integrity_status": "STORAGE_VERIFIED",
                  "status": "UPLOADED"
                },
                "schema": {
                  "$ref": "#/components/schemas/UploadCompleteResponse"
                }
              }
            },
            "description": "Upload verified"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Complete a video upload",
        "tags": [
          "Uploads"
        ]
      }
    },
...
  },
...
}